GHOST-ZCS: actions sind nur noch INTs # pro ghost ein observer+classifierSystem # einige grundlegende bewegungsregeln (noch zu testen ..) # matchset wird selektiert # actionset selektiert diejenige action mit hoechster gesamtsumme (selektionen s. debug-ausgabe im game)

+++ PacMan/src/classifier/IObservation.java
@@ -1,8 +1,14 @@
 package classifier;
 
 public interface IObservation {
+	
+	//
+	// currently based on binary kram
+	//
+	
+	
 	boolean matches(IObservation other);
-	boolean matches(long observationBits);
-	long getObservedConditions(); // bit=0 --> false, bit=1 --> true # max 64 conditions bei long-typ
+	boolean matches(int observationBits);
+	int getObservedConditions(); // bit=0 --> false, bit=1 --> true # max 64 conditions bei long-typ
 	//long getWildcards(); // bit=0 --> no wildcard, bit=1 --> isWildcard (ONLY FOR DATABASE USE)
 }
